fix: move back URLUtils test
